+++ .storybook/main.ts
@@ -6,10 +6,9 @@ import {
   nodeNextExtensionAlias,
 } from "wo-library/tools/storybook/webpack";
 
-main.addons.push("@storybook/addon-styling");
-
 const config: StorybookConfig = {
   ...main,
+  addons: [...(main.addons || []), "@storybook/addon-styling"],
   webpackFinal: (config, { configType }) => {
     // `configType` has a value of 'DEVELOPMENT' or 'PRODUCTION'
     // 'PRODUCTION' is used when building the static version of storybook.