+++ .storybook/preview.ts
@@ -4,7 +4,13 @@ import "../src/styles/base.css";
 import "./preview.css";
 
 const preview: Preview = {
-  parameters: {},
+  parameters: {
+    // for addon-actions to work with regex with TS
+    // add argTypes explicitly for components which extend
+    // the change handlers from a different type till this issue is resolved
+    // https://github.com/storybookjs/storybook/issues/15012
+    actions: { argTypesRegex: "^on.*" },
+  },
 };
 
 export default preview;